+++ 5.7
@@ -1,0 +1,3 @@
+2022-12-06 22:17:38+00:00 [Note] [Entrypoint]: Entrypoint script for MySQL Server 8.0.31-1.el8 started.
+2022-12-06 22:17:39+00:00 [Note] [Entrypoint]: Switching to dedicated user 'mysql'
+2022-12-06 22:17:39+00:00 [Note] [Entrypoint]: Entrypoint script for MySQL Server 8.0.31-1.el8 started.
