+++ Dockerfile
@@ -1,0 +1,5 @@
+FROM openjdk:11-jdk-alpine
+ARG JAR_FILE=target/*.jar
+COPY ${JAR_FILE} app.jar
+ENTRYPOINT ["java","-jar","/app.jar"]
+EXPOSE 8080
