+++ docker-compose.yml
@@ -1,35 +1,21 @@
-version: "3.7"
+version: '3.7'
 
 services:
-
-##  mongo:
-##    image: mongo:4.2
-##    hostname: mongo
-##    container_name: mongo
-##    ports:
-##      - 27017:27017
-##    environment:
-##      MONGO_INITDB_DATABASE: miniautorizador
-##      MONGO_INITDB_ROOT_USERNAME: user
-##      MONGO_INITDB_ROOT_PASSWORD: password
-##   volumes:
-##      - ./scripts/init.js:/docker-entrypoint-initdb.d/init.js:ro
-
-  mysql:
-    image: mysql:5.7
-    hostname: mysql
-    container_name: mysql
-    restart: always
-    ports:
-      - "3306:3306"
-    environment:
-      MYSQL_DATABASE: miniautorizador
-      MYSQL_ROOT_PASSWORD: root
-      MYSQL_ALLOW_EMPTY_PASSWORD: yes
-    
-##   api:
-##    build: "./api"
-##   container_name: node-container
-##    restart: always
-##     ports: 
-##        - "8080:8080"  
+#    autorizadorvr:
+#      image: autorizadorvr:0.0.1-SNAPSHOT
+#      restart: always
+#      ports:
+#        - 8080:8080
+#      links:
+#        - mysql
+      mysql:
+      image: mysql:5.7
+      hostname: mysql
+      container_name: mysql
+      restart: always
+      ports:
+       - 3306:3306
+      environment:
+        MYSQL_DATABASE: miniautorizador
+        MYSQL_ROOT_PASSWORD: root
+        MYSQL_ALLOW_EMPTY_PASSWORD: 'yes'
