--- src/main/java/br/com/api/autorizadorVR/AutorizadorVrApplication.java
@@ -1,15 +1,0 @@
-package br.com.api.autorizadorVR;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-
-@SpringBootApplication
-public class AutorizadorVrApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AutorizadorVrApplication.class, args);
-		
-	}
-
-}
