--- src/main/java/br/com/api/autorizadorVR/config/ModelMapperConfig.java
@@ -1,16 +1,0 @@
-package br.com.api.autorizadorVR.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ModelMapperConfig {
-
-	
-	@Bean
-	public ModelMapper modelMapper() {
-		
-		return new ModelMapper();
-	}
-}
