+++ src/main/java/br/com/api/autorizadorvr/AutorizadorVrApplication.java
@@ -1,0 +1,15 @@
+package br.com.api.autorizadorvr;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+
+@SpringBootApplication
+public class AutorizadorVrApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AutorizadorVrApplication.class, args);
+		
+	}
+
+}
