+++ src/main/java/br/com/api/autorizadorvr/config/DataBaseConfig.java
@@ -1,0 +1,41 @@
+package br.com.api.autorizadorvr.config;
+
+import javax.sql.DataSource;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+import org.springframework.orm.jpa.JpaVendorAdapter;
+import org.springframework.orm.jpa.vendor.Database;
+import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
+
+@Configuration
+public class DataBaseConfig {
+
+	@Bean
+	public DataSource getDataSource() {
+
+		DriverManagerDataSource dataSource = new DriverManagerDataSource();
+		 
+		dataSource.setDriverClassName("com.mysql.jdbc.Driver");
+		dataSource.setUrl("jdbc:mysql://localhost:3306/miniautorizador?useTimezone=true&serverTimezone=UTC");
+		dataSource.setUsername("root");
+		dataSource.setPassword("root");
+		
+		return dataSource;
+	}
+
+	@Bean
+	public JpaVendorAdapter jpaVendorAdapter() {
+		HibernateJpaVendorAdapter adapter = new HibernateJpaVendorAdapter();
+		adapter.setDatabase(Database.MYSQL);
+		adapter.setShowSql(true);
+		adapter.setGenerateDdl(true);
+		
+		adapter.setDatabasePlatform("org.hibernate.dialect.MySQLDialect");
+		adapter.setShowSql(true);
+		adapter.setPrepareConnection(true);
+		return adapter;
+	}
+	
+}
