+++ src/main/java/br/com/api/autorizadorvr/config/ModelMapperConfig.java
@@ -1,0 +1,16 @@
+package br.com.api.autorizadorvr.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ModelMapperConfig {
+
+	
+	@Bean
+	public ModelMapper modelMapper() {
+		
+		return new ModelMapper();
+	}
+}
