+++ src/main/java/br/com/api/autorizadorvr/dto/SaldoDTO.java
@@ -1,0 +1,29 @@
+package br.com.api.autorizadorvr.dto;
+
+import java.io.Serializable;
+
+import javax.persistence.Id;
+import javax.validation.constraints.NotBlank;
+
+import org.hibernate.type.BigIntegerType;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class SaldoDTO implements Serializable{
+
+	private static final long serialVersionUID = 1L;
+	
+	private Long saldo;
+
+}
