+++ src/main/java/br/com/api/autorizadorvr/exception/InternalServerErrorException.java
@@ -1,0 +1,31 @@
+package br.com.api.autorizadorvr.exception;
+
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value= HttpStatus.INTERNAL_SERVER_ERROR)
+public class InternalServerErrorException extends RuntimeException{
+
+	private static final long serialVersionUID = 1L;
+
+	public InternalServerErrorException() {
+		super();
+	}
+	
+
+	public InternalServerErrorException(String message) {
+		super(message);
+	}
+
+	public InternalServerErrorException(Throwable cause) {
+		super(cause);
+	}
+
+
+	public InternalServerErrorException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+}
+
