+++ src/main/resources/application.properties
@@ -1,2 +1,3 @@
+spring.profiles.active=dev
 spring.mvc.pathmatch.matching-strategy=ant-path-matcher
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL57Dialect
