+++ src/test/java/br/com/api/autorizadorVR/AutorizadorVrApplicationTests.java
@@ -3,10 +3,10 @@ package br.com.api.autorizadorVR;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+
 class AutorizadorVrApplicationTests {
 
-	@Test
+
 	void contextLoads() {
 	}
